requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.4
tzdata==2025.3
urllib3==2.6.2
